colours will have a much lower chance of being the same next to each other
although its not perfect, it's good enough. also, we messed around with the graph styling a bit, and there are three modes defined in the code, so that all the colours can be the same, just the graphs, or what it was before. Since it's a bit too hard to manage states with the non async graphs, it hasn't yet been implemented into the app, but it exists in the code if we ever want to implement it in the future